Make own interval loop to handle pausing
Instead of using the built-in setInterval, we implement our own interval loop,
which receives tick(deltaTime) calls. This allows intervals to not elapse when
the game is paused.

This is very important for continuous spawns to not break, otherwise merely
pausing the game can result in a huge spike of spawns.