Bump prototype version to v0.0.3
I am getting quite satisfied with the recent changes. From adding new
projectile types, to merchant loot, to automatic inventory management, and
much more, as well as a slurry of bug fixes and minor improvements and
touch-ups and whatnot, I think we have achieved a bit of a milestone in Loot &
Roam's history here.